Incorporated The core of the Upgrades system
Some variables will need to be moved to a centralized location, but we'll worry about that later.
Edited some of the button names for the widgets, and created an "Upgrades" button that opens this:

Clicking "Close" in the upgrades menu, or anywhere outside the white box closes the upgrades menu.

"Better Click" is coded, but I need to find a better way to reference the variables in UI_HudTest1 (Due to this, it currently does not function). We likely need to move the variables to a more centralized location so we don't have to inherit the values from one specific UI Blueprint, though that method is escaping my mind atm.
*when functional, numbers on the left should show how many times each upgrade has been purchased
& numbers on the right showcase the cost, which increases multiplicatively.
Outside of the broken reference, Clicking the "Better Click" button would check to see if the user has enough currency to buy the upgrade, then deduct the cost from their currency pool, and multiply the cost of the next upgrade by 1.5x, and add to the Note Earn Multiplier I set up in the other blueprint (also increment the displayed "Upgrade1Count"