Rudimentary Rhythmic Timing Bonus
Timing bonus: It works using the framework that Braden built with the Metrognome file. I created a bunch of Min and Max values in both the Metrognome, and BaseGame_Mode files, that way EventTick can keep track of the bar's position, and provide an appropriate (currently temporarily set) bonus according to when you tap.

TL;DR, if you don't have a massive skill issue like me, hitting the Sweet (Green) spot awards 100 notes, and puts the bonus on a 0.04s Cooldown (to prevent spam).
Hitting the Sour (Yellow) spot awards 50 notes, and puts the bonus on a 1s Cooldown
Missing (Hitting the Red spot) awards no bonus, and puts the bonus on a 2s Cooldown