Updates to the Rhythmic Timing Bonus, hooked up SUpgrade#2
Set up the 2nd Upgrade so the timing Bonus can actually be tuned with those values. Playing a lot of Idle Universe made me realize that I could theoretically tap ~20 times/second, so I made the base multiplier of the bonus effectively 30x the normal tap rate (scales linearly with Upgrade 1) to provide an incentive to tap to the beat.
Updated the "Cooldown" of the Rhythmic Timing bonus to keep resetting the Cooldown if a tap happened while on-cooldown (This makes it much harder to "accidentally" get the bonus while spamming the tap button)

Self Upgrade 3, based on Trello, can be worked on next, and should take the size and dimension values of the yellow and green boxes that I turned into variables, and increase them as a way to increase the leniency of the Timing bonus (This upgrade is expensive to start, currently set to 1.5k Notes, and would increase its cost exponentially, so as to appear more like a luxury rather than a requirement). I'm thinking a good place to start would be to make the green box 1 or 2 pixels larger per purchase of Upgrade 3, with the Yellow box being 2 or 4 pixels larger, depending on what's chosen, with this upgrade having a max amount of purchases when the yellow box is the same width as the red box (i.e, when you literally CAN'T miss the sour spot). Doing the math with the current layout, Green at 1 pixel with Yellow at 2 would make the max upgrades be 250, with G2Y4's max upgrades being 125. So... whichever one works better in your mind to start, @theappleholic

We still need to sync with a general rhythm, which will take tuning, but that balance can come later.