Bugfixes, and UI Updates
Fixing some pretty major tech bugs, decided to re-shuffle some of the UI + add some polish to the tap button
It now rotates a full 360 passively over 30 seconds if untouched. When a tap button is pressed, (I can make it be when JUST tapping the record instead if we want), then the rotation briefly accelerates.

I'm potentially noticing that the music loop may desync every *other* loop. Leads me to believe that the loop timing is somehow off. No other elements get out of sync, the metronome remains consistent, the spinning remains consistent. Still gotta do some research to figure ***that*** out.
Also, to add to the above, Self Upgrade 2 WAS functioning, but it was statistically bad, so I buffed it to 1.15x instead of 1.05x, and nerfed the cost to 50 instead of 20.
Updated Current UI art to not look so damn blurry in lieu of getting higher-res assets (how-to in attached image)

Next steps now include implementing the teaser feature of the Instrument Upgrades. For now, I've been theorycrafting how to ACTUALLY do that.