First Pass Venue 1 -- Cloud Background Anim
First Pass for a cloud ambient anim is likely not ideal for timing purposes. However, it is one of the easiest dev-side to implement, as there's no "movement" to the actual images -- they're just sliding along the screen.

For the background/foreground, if we want the effect of the clouds passing behind the coastline, the background will also need to be exported separately. Making both assets the exact same size has a strange, and noticeable effect on the asset placement/size. It's a bit squished and not perfect, definitely easily seen on the lighthouse itself.

What I anticipate happening now, is putting each animation in a separate UI BP, then we can leverage a bit more of a robust system for showcasing level assets to then call in the correct elements at the correct intervals.

Still working on iterating on how I'd want the clouds in the scene to move, especially to the rhythm, but hopefully this could give a solid idea of how to move forward in the meantime.